Performance improved
I improved the performance of the program. It is much faster than the last version, but it is still slower than Matlab. The performance may be half of Matlab. I found some problem. If the force condition is changed, the code also should be changed. The input doesn’t include the information of forces. This code is only applicable to the input (60, 20, 4, 0.3, 3, 1.5) and (60, 20, 4, 0.3, 3, 1.5).